Testing rc is not enough
In most of the case rc will almost always return 0. Testing the return
code of the curl command simply makes sure that the command has been
well executed, it doesn't mean that the request worked properly.
For instance rc could return 0 but the curl could have gotten a 401
code. This is why we test both, rc and http return codes.