heat-engine: disable the check on database connection
After some investigations, the mysql connection is not persistent and is
only established when heat API is used by a client.
By checking if the TCP connection is actually established, we have a fail
every time Pacemaker checks. If no Heat API requests have been sent for
some time, Heat will be considered as failed and will be restart which
is not clean.

This patch aims to disable this check and only let AMQP connection
check.

Signed-off-by: Emilien Macchi <emilien.macchi@enovance.com>